Refactor salt-models testing
The tests take too much time to be executed and they often fail due to lack
of resources. The solution to improve salt models testing is to disitribute
the work among Jenkins slave. The previous approach consisted in doing all the
tests on one slave for a given model.
Fix test for system-reclass by testing the good reclass-system version.

Change-Id: I605355bdb66d7195e42454bb87b5cc0663d13929