Fix uploading OpenScap results to Worp API
sendHttpRequest function fails on any response code other than 200,
which is in fact a pretty standard case for a REST API, including Worp.

Re-use the restCall (and its derivatives) method instead.

Also do not create per-node reports with single record,
instead push all nodes as records to a single report.

In preparation for CVE scanning, also make the method that uploads
to worp take a report type as input.

Change-Id: I508a31a0e461869587213778465a88585487977f