Allow to set custom STACK_NAME when creating infra
This patch allows to create custom STACK_NAME when crating infra but
user-id will still be added to make sure we save check for max allowed
stack per user.

This is needed to avoid issues when paren job is terminated, and cleanup
failed because stack-name is not available to parent job at this time.

Change-Id: I834cfc22ad631d71b3af4587195192dd38e400ac
Related-Prod: PROD-19102